[REF] - S#13114 - Move security from coop_membership to coop_membership_access_right
