+++ intercoop_addons/coop_membership/views/view_res_partner.xml
@@ -103,14 +103,14 @@
                     <group>
                         <group string="MemberShip">
                             <field name="barcode_base" string="Member Number" readonly="1"/>
-                            <field name="is_louve_member" groups="coop_membership.coop_membership_manager" string="Afficher via menu Membres" readonly="0"/>
-                            <field name="type_A_capital_qty" groups="coop_membership.coop_membership_manager"/>
-                            <field name="type_B_capital_qty" groups="coop_membership.coop_membership_manager"/>
-                            <field name="type_C_capital_qty" groups="coop_membership.coop_membership_manager"/>
+                            <field name="is_louve_member" string="Afficher via menu Membres" readonly="0"/>
+                            <field name="type_A_capital_qty"/>
+                            <field name="type_B_capital_qty"/>
+                            <field name="type_C_capital_qty"/>
                             <field name="is_type_A_capital_subscriptor" invisible="1"/>
                             <field name="is_type_B_capital_subscriptor" invisible="1"/>
                             <field name="is_type_C_capital_subscriptor" invisible="1"/>
-                            <field name="is_associated_people" readonly="1" groups="coop_membership.coop_membership_manager"/>
+                            <field name="is_associated_people" readonly="1"/>
                             <field name="is_unpayed"/>
                             <field name="is_unsubscribed"/>
                         </group>
@@ -151,12 +151,6 @@
             <field name="fundraising_partner_type_ids" position="attributes">
                 <attribute name="readonly">1</attribute>
             </field>
-            <page name="page_capital_fundraising" position="attributes">
-                <attribute name="groups" ref="coop_membership.subscriptions_reader" />
-            </page>
-            <xpath expr="//button[@id='new_subscription']" position="attributes">
-                <attribute name="groups" ref="coop_membership.subscriptions_creator" />
-            </xpath>
             <field name="fundraising_partner_type_ids" position="after">
                 <button name="set_underclass_population" type="object"
                     string="Set 'Underclass Population'" colspan="2" class="oe_highlight" attrs="{'invisible': [('is_underclass_population', '=', True)]}"/>
@@ -171,7 +165,6 @@
     <record model="ir.ui.view" id="view_res_partner_form_subscription_manager">
         <field name="model">res.partner</field>
         <field name="inherit_id" ref="view_res_partner_form_subscription" />
-        <field name="groups_id" eval="[(6,0,[ref('account.group_account_manager')])]" />
         <field name="arch" type="xml">
             <xpath expr="//field[@name='fundraising_partner_type_ids']" position="attributes">
                 <attribute name="readonly">0</attribute>
