+++ louve_addons/coop_membership_access_rights/__init__.py
@@ -21,3 +21,4 @@
 ##############################################################################
 
 import models
+import wizard
