+++ louve_addons/coop_membership_access_rights/__openerp__.py
@@ -55,11 +55,14 @@
         'stock',
     ],
     'data': [
+        'security/ir_module_category.xml',
         'security/res_group.xml',
         'security/ir_model_access.yml',
+        'security/ir.model.access.csv',
         'views/menu.xml',
         'views/res_partner_view.xml',
         'views/res_users_view.xml',
         'views/account_invoice_view.xml',
+        'wizard/view_capital_fundraising_wizard.xml',
     ],
 }
