+++ louve_addons/coop_membership_access_rights/security/ir.model.access.csv
@@ -1,0 +1,31 @@
+id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
+,,,,,,,
+,,,,,,,
+access_model_capital_fundraising_member,access_model_capital_fundraising_member,capital_subscription.model_capital_fundraising,coop_membership_access_rights.subscriptions_reader,1,0,0,0
+access_model_capital_fundraising_category_member,access_model_capital_fundraising_category_member,capital_subscription.model_capital_fundraising_category,coop_membership_access_rights.subscriptions_reader,1,0,0,0
+access_model_capital_fundraising_category_line_member,access_model_capital_fundraising_category_line_member,capital_subscription.model_capital_fundraising_category_line,coop_membership_access_rights.subscriptions_reader,1,0,0,0
+access_model_capital_fundraising_partner_type_member,access_model_capital_fundraising_partner_type_member,capital_subscription.model_capital_fundraising_partner_type,coop_membership_access_rights.subscriptions_reader,1,0,0,0
+,,,,,,,
+,,,,,,,
+,,,,,,,
+access_model_account_invoice_member,access_model_account_invoice_member,account.model_account_invoice,coop_membership_access_rights.subscriptions_reader,1,0,0,0
+access_model_account_invoice_line_member,access_model_account_invoice_line_member,account.model_account_invoice_line,coop_membership_access_rights.subscriptions_reader,1,0,0,0
+access_model_account_journal_member,access_model_account_journal_member,account.model_account_journal,coop_membership_access_rights.subscriptions_reader,1,0,0,0
+access_model_account_invoice_tax_member,access_model_account_invoice_tax_member,account.model_account_invoice_tax,coop_membership_access_rights.subscriptions_reader,1,0,0,0
+access_model_account_move_line_member,access_model_account_move_line_member,account.model_account_move_line,coop_membership_access_rights.subscriptions_reader,1,0,0,0
+access_model_account_partial_reconcile_member,access_model_account_partial_reconcile_member,account.model_account_partial_reconcile,coop_membership_access_rights.subscriptions_reader,1,0,0,0
+access_model_account_account_type_member,access_model_account_account_type_member,account.model_account_account_type,coop_membership_access_rights.subscriptions_reader,1,0,0,0
+access_model_account_analytic_line_member,access_model_account_analytic_line_member,account.model_account_analytic_line,coop_membership_access_rights.subscriptions_reader,1,0,0,0
+access_model_account_payment_member,access_model_account_payment_member,account.model_account_payment,coop_membership_access_rights.subscriptions_reader,1,0,0,0
+,,,,,,,
+,,,,,,,
+access_model_account_invoice_creator,access_model_account_invoice_creator,account.model_account_invoice,coop_membership_access_rights.subscriptions_creator,1,1,1,0
+access_model_account_invoice_line_creator,access_model_account_invoice_line_creator,account.model_account_invoice_line,coop_membership_access_rights.subscriptions_creator,1,1,1,0
+access_model_account_journal_creator,access_model_account_journal_creator,account.model_account_journal,coop_membership_access_rights.subscriptions_creator,1,1,1,0
+access_model_account_invoice_tax_creator,access_model_account_invoice_tax_creator,account.model_account_invoice_tax,coop_membership_access_rights.subscriptions_creator,1,1,1,0
+access_model_account_move_creator,access_model_account_move_creator,account.model_account_move,coop_membership_access_rights.subscriptions_creator,1,1,1,0
+access_model_account_move_line_creator,access_model_account_move_line_creator,account.model_account_move_line,coop_membership_access_rights.subscriptions_creator,1,1,1,0
+,,,,,,,
+access_model_account_payment_creator_payment,access_model_account_payment_creator_payment,account.model_account_payment,coop_membership_access_rights.subscriptions_creator_payment,1,1,1,0
+,,,,,,,
+access_model_res_contact_origin,access_model_res_contact_origin,coop_membership.model_res_contact_origin,,1,1,1,1
