+++ louve_addons/coop_membership_access_rights/security/ir_module_category.xml
@@ -6,9 +6,7 @@
 -->
 
 <odoo>
-
     <record model="ir.module.category" id="category_subscriptions">
         <field name="name">Subscriptions</field>
     </record>
-
 </odoo>
