+++ louve_addons/coop_membership_access_rights/security/res_group.xml
@@ -45,4 +45,33 @@
         <field name="category_id" ref="module_membership_access_management"/>
         <field name="users" eval="[(4, ref('base.user_root'))]"/>
     </record>
+
+
+    <record model="res.groups" id="subscriptions_reader">
+        <field name="name">See Subscriptions</field>
+        <field name="category_id" ref="category_subscriptions" />
+    </record>
+
+    <record model="res.groups" id="subscriptions_creator">
+        <field name="name">Create Subscriptions</field>
+        <field name="category_id" ref="category_subscriptions" />
+        <field name="implied_ids" eval="[(4, ref('coop_membership_access_rights.subscriptions_reader'))]" />
+    </record>
+
+    <record model="res.groups" id="subscriptions_creator_payment">
+        <field name="name">Create Subscriptions and payment</field>
+        <field name="category_id" ref="category_subscriptions" />
+        <field name="implied_ids" eval="[(4, ref('coop_membership_access_rights.subscriptions_creator'))]" />
+    </record>
+
+    <record model="res.groups" id="subscriptions_can_change_fundraising_category">
+        <field name="name">Can Change Fundraising Category (in Wizard)</field>
+        <field name="category_id" ref="base.module_category_extra" />
+    </record>
+
+    <record model="res.groups" id="coop_membership_manager">
+        <field name="name">Manage Coop Membership (Full)</field>
+        <field name="category_id" ref="base.module_category_extra" />
+    </record>
+
 </odoo>
