+++ louve_addons/coop_membership_access_rights/wizard/capital_fundraising_wizard.py
@@ -1,0 +1,19 @@
+# -*- coding: utf-8 -*-
+# Copyright (C) 2016-Today: La Louve (<http://www.lalouve.net/>)
+# @author: Sylvain LE GAL (https://twitter.com/legalsylvain)
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+
+from openerp import models, fields, api
+
+
+class CapitalFundraisingWizard(models.TransientModel):
+    _inherit = 'capital.fundraising.wizard'
+
+    def default_can_change_fundraising_category(self):
+        return self.user_has_groups(
+            'coop_membership_access_rights.' +
+            'subscriptions_can_change_fundraising_category')
+
+    can_change_fundraising_category = fields.Boolean(
+        string="Can Change fundraising Category",
+        default=default_can_change_fundraising_category)
