+++ louve_addons/coop_membership_access_rights/wizard/view_capital_fundraising_wizard.xml
